109-tencet-ModifySecurityGroupPolicies/pyproject.toml
[tool.poetry]
name = "109-tencet-modifysecuritygrouppolicies"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.6"
tencentcloud-sdk-python = "^3.0.683"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
